Make all the processes terminate properly.

+++ laba3/main.cpp
@@ -74,16 +74,17 @@ std::optional<std::string> iter_bytes(std::string pswd, std::string hash,
 // main process overload
 std::optional<std::string> iter_bytes(std::string pswd, std::string hash,
                                       std::string stop_word, int pids[3]) {
+  for (int i = 0; i < 3; i++) {
+    std::cout << "Recieved pid is: " << pids[i] << std::endl;
+  }
   int i = 5;
   while (pswd != stop_word) {
     for (int i = 0; i < 3; i++) {
-      if (kill(pids[i], 0) != 0) {
+      if (getpgid(pids[i] < 0)) {
+        std::cout << "Process with pid " << pids[i] << " has finished earlier" << std::endl;
         for (int j = 0; j < 3; j++) {
-          if (j != i) {
-            kill(pids[j], SIGTERM);
-          }
+            if (j != i) kill(pids[j], SIGTERM);
         }
-        MPI_Finalize();
         exit(0);
       }
     }
@@ -155,6 +156,9 @@ int main(int argc, char *argv[]) {
         if (res) {
           std::cout << "Password found: " << *res << std::endl;
         }
+
+        std::cout << "Process " << current << " found the password" << std::endl;
+        MPI_Finalize();
       }
     }
   }
